DynaHelper
A macro that is designed to be interfaced with by a DynaVox system.  It
records and organizes basic computer input events and  replays them at a
future time upon the single selection of a DynaVox button.  The DynaVox
communicates with the DynaHelper (which is on a separate Windows PC) via
AccessIt.  This program improves the speed of use of common and simple
computer applications for individuals with severe physical handicaps.